% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negbin_model.R
\name{grad_log_posterior_negbin_log_r}
\alias{grad_log_posterior_negbin_log_r}
\title{Gradient of the Log-posterior for Negative Binomial Regression with Log-Overdispersion}
\usage{
grad_log_posterior_negbin_log_r(
  theta,
  X,
  y,
  beta_mu,
  beta_sigma,
  alpha_r,
  beta_r
)
}
\arguments{
\item{theta}{A numeric vector containing regression coefficients (beta) and the
log-dispersion parameter (phi), i.e., theta = c(beta, phi).}

\item{X}{A numeric matrix of predictors.}

\item{y}{A numeric vector of observed counts.}

\item{beta_mu}{A numeric vector of prior means for the regression coefficients.}

\item{beta_sigma}{A numeric vector or scalar representing the prior standard
deviations for the regression coefficients.}

\item{alpha_r}{A numeric scalar representing the shape parameter of the gamma
prior for r = exp(phi).}

\item{beta_r}{A numeric scalar representing the rate parameter of the gamma
prior for r = exp(phi).}
}
\value{
A numeric vector representing the gradient of the log-posterior with
  respect to theta = c(beta, phi).
}
\description{
This function computes the gradient of the log-posterior with respect to the regression
coefficients (beta) and the log-dispersion parameter (phi = log(r)).
}
