% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negbin_model.R
\name{log_posterior_negbin_log_r}
\alias{log_posterior_negbin_log_r}
\title{Log-posterior for Negative Binomial Regression with Log-Overdispersion Parameter}
\usage{
log_posterior_negbin_log_r(theta, X, y, beta_mu, beta_sigma, alpha_r, beta_r)
}
\arguments{
\item{theta}{A numeric vector containing regression coefficients (beta) and the
log-dispersion parameter (phi), i.e., theta = c(beta, phi).}

\item{X}{A numeric matrix of predictors.}

\item{y}{A numeric vector of response counts.}

\item{beta_mu}{A numeric vector of prior means for the regression coefficients.}

\item{beta_sigma}{A numeric vector or scalar representing the prior standard
deviations for the regression coefficients.}

\item{alpha_r}{A numeric scalar representing the shape parameter of the gamma
prior for r = exp(phi).}

\item{beta_r}{A numeric scalar representing the rate parameter of the gamma
prior for r = exp(phi).}
}
\value{
A single numeric value representing the log-posterior.
}
\description{
This function calculates the log-posterior for a Negative Binomial regression model,
where the overdispersion parameter is parameterized as phi = log(r).
}
